requirements.txt
Django==4.0.6
Pillow==9.2.0
django-leaflet==0.28.3
django-crispy-forms==1.14.0
crispy-bootstrap5==0.7